+++ internal/processhandle/signal_other.go
@@ -10,3 +10,5 @@ func openHandle(int) (*handle, error)       { return nil, ErrUnsupported }
 func (*handle) close()                      {}
 func (*handle) alive() error                { return ErrUnsupported }
 func (*handle) signal(syscall.Signal) error { return ErrUnsupported }
+
+func probe() error { return ErrUnsupported }